// cluster/console/console/web/package/src/apis/cordiumv1/cordiumv1.ts
export interface Workspace {
    /**
     * APIVersion is the API version (i.e. "cordium/v1")
     *
     * @generated from protobuf field: string apiVersion = 1
     */
    apiVersion: string;
    /**
     * Kind is the resource name (i.e. `Workspace`).
     *
     * @generated from protobuf field: string kind = 2
     */
    kind: string;
    /**
     * Metadata is the object's metadata.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.Metadata metadata = 3
     */
    metadata?: Metadata;
    /**
     * Spec is the Workspace specification.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Spec spec = 4
     */
    spec?: Workspace_Spec;
    /**
     * Status is the current status of the Workspace.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.Workspace.Status status = 5
     */
    status?: Workspace_Status;
}
// ...
export interface Workspace_Spec_Var {
    /**
     * Name is the variable's name.
     *
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * Value is the variable's value.
     *
     * @generated from protobuf field: string value = 2
     */
    value: string;
}
// ...
export enum Workspace_Status_State {
    /**
     * UNKNOWN is not used.
     *
     * @generated from protobuf enum value: UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * INIT_REQUEST means that a start request has been accepted by the API
     * server.
     *
     * @generated from protobuf enum value: INIT_REQUEST = 1;
     */
    INIT_REQUEST = 1,
    /**
     * INITIALIZING means that the Cluster is provisioning the Workspace's
     * underlying resources and waiting for its supervisor to become ready.
     *
     * @generated from protobuf enum value: INITIALIZING = 2;
     */
    INITIALIZING = 2,
    /**
     * PULLING_IMAGE means that the container image is being pulled from the
     * registry.
     *
     * @generated from protobuf enum value: PULLING_IMAGE = 3;
     */
    PULLING_IMAGE = 3,
    /**
     * BUILDING_IMAGE means that the container image is being built from a
     * Dockerfile or a devcontainer spec.
     *
     * @generated from protobuf enum value: BUILDING_IMAGE = 4;
     */
    BUILDING_IMAGE = 4,
    /**
     * STARTING_RUNTIME means that the Workspace's container has started and
     * that the agent is initializing.
     *
     * @generated from protobuf enum value: STARTING_RUNTIME = 5;
     */
    STARTING_RUNTIME = 5,
    /**
     * PREPARING means that the agent is running the lifecycle setup (i.e.
     * the repository cloning, the ON_CREATE tasks, the dotfiles and the
     * devcontainer Features).
     *
     * @generated from protobuf enum value: PREPARING = 6;
     */
    PREPARING = 6,
    /**
     * RUNNING means that the Workspace is fully initialized and ready to be
     * used.
     *
     * @generated from protobuf enum value: RUNNING = 7;
     */
    RUNNING = 7,
    /**
     * STOPPING_REQUEST means that a stop request has been received.
     *
     * @generated from protobuf enum value: STOPPING_REQUEST = 9;
     */
    STOPPING_REQUEST = 9,
    /**
     * STOPPING means that the Workspace is shutting down gracefully and
     * running its PRE_STOP tasks.
     *
     * @generated from protobuf enum value: STOPPING = 10;
     */
    STOPPING = 10,
    /**
     * STOPPED means that the Workspace is not running. Its storage is
     * preserved unless the Workspace is ephemeral.
     *
     * @generated from protobuf enum value: STOPPED = 11;
     */
    STOPPED = 11
}
// ...
export interface StartWorkspaceRequest {
    /**
     * WorkspaceRef is the reference of the Workspace to be started.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference workspaceRef = 1
     */
    workspaceRef?: ObjectReference;
    /**
     * Config is the run-specific configuration of the run being started.
     *
     * @generated from protobuf field: octelium.api.main.cordium.v1.StartWorkspaceRequest.Config config = 3
     */
    config?: StartWorkspaceRequest_Config;
}
/**
 * Config is the run-specific configuration that only applies to the run
 * being started.
 *
 * @generated from protobuf message octelium.api.main.cordium.v1.StartWorkspaceRequest.Config
 */
export interface StartWorkspaceRequest_Config {
    /**
     * Vars is the list of the variables that override the Workspace's own
     * variables for this run.
     *
     * @generated from protobuf field: repeated octelium.api.main.cordium.v1.Workspace.Spec.Var vars = 1
     */
    vars: Workspace_Spec_Var[];
    /**
     * RegionRef is the reference of the Region that is chosen to host the run.
     * If it is unset, the Cluster picks a Region on its own, preferring the
     * User's preferred Region if it is set in their UserConfig.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference regionRef = 2
     */
    regionRef?: ObjectReference;
}
// ...
export interface StopWorkspaceRequest {
    /**
     * WorkspaceRef is the reference of the Workspace to be stopped.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ObjectReference workspaceRef = 1
     */
    workspaceRef?: ObjectReference;
}
// ...
export const Workspace = new Workspace$Type();
// ...
export const Workspace_Spec = new Workspace_Spec$Type();
// ...
export const Workspace_Spec_Var = new Workspace_Spec_Var$Type();
// ...
export const Workspace_Status = new Workspace_Status$Type();
// ...
class StartWorkspaceRequest$Type extends MessageType<StartWorkspaceRequest> {
    constructor() {
        super("octelium.api.main.cordium.v1.StartWorkspaceRequest", [
            { no: 1, name: "workspaceRef", kind: "message", T: () => ObjectReference },
            { no: 3, name: "config", kind: "message", T: () => StartWorkspaceRequest_Config }
        ]);
    }
    create(value?: PartialMessage<StartWorkspaceRequest>): StartWorkspaceRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<StartWorkspaceRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: StartWorkspaceRequest): StartWorkspaceRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.ObjectReference workspaceRef */ 1:
                    message.workspaceRef = ObjectReference.internalBinaryRead(reader, reader.uint32(), options, message.workspaceRef);
                    break;
                case /* octelium.api.main.cordium.v1.StartWorkspaceRequest.Config config */ 3:
                    message.config = StartWorkspaceRequest_Config.internalBinaryRead(reader, reader.uint32(), options, message.config);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: StartWorkspaceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.ObjectReference workspaceRef = 1; */
        if (message.workspaceRef)
            ObjectReference.internalBinaryWrite(message.workspaceRef, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* octelium.api.main.cordium.v1.StartWorkspaceRequest.Config config = 3; */
        if (message.config)
            StartWorkspaceRequest_Config.internalBinaryWrite(message.config, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.cordium.v1.StartWorkspaceRequest
 */
export const StartWorkspaceRequest = new StartWorkspaceRequest$Type();
// ...
export const StartWorkspaceRequest_Config = new StartWorkspaceRequest_Config$Type();
// ...
class StopWorkspaceRequest$Type extends MessageType<StopWorkspaceRequest> {
    constructor() {
        super("octelium.api.main.cordium.v1.StopWorkspaceRequest", [
            { no: 1, name: "workspaceRef", kind: "message", T: () => ObjectReference }
        ]);
    }
    create(value?: PartialMessage<StopWorkspaceRequest>): StopWorkspaceRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<StopWorkspaceRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: StopWorkspaceRequest): StopWorkspaceRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.ObjectReference workspaceRef */ 1:
                    message.workspaceRef = ObjectReference.internalBinaryRead(reader, reader.uint32(), options, message.workspaceRef);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: StopWorkspaceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.ObjectReference workspaceRef = 1; */
        if (message.workspaceRef)
            ObjectReference.internalBinaryWrite(message.workspaceRef, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.cordium.v1.StopWorkspaceRequest
 */
export const StopWorkspaceRequest = new StopWorkspaceRequest$Type();